requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
httpx>=0.27.0
redis>=5.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-instrumentation-fastapi>=0.44b0
tiktoken>=0.6.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
anyio>=4.0.0
